Update to date

+++ app/src/main/java/vn/vistark/stm/data/model/ClassObj.kt
@@ -1,11 +1,11 @@
 package vn.vistark.stm.data.model
 
 class ClassObj() {
-    var id: Int = -1
+    var id: Long = -1
     var name: String = ""
     var description: String = ""
 
-    constructor(id: Int, name: String, description: String) : this() {
+    constructor(id: Long, name: String, description: String) : this() {
         this.id = id
         this.name = name
         this.description = description
